requirements.txt
Flask==3.0.3
python-telegram-bot==20.7
Telethon==1.34.0
python-dotenv==1.0.1
cryptography==42.0.5
APScheduler==3.10.4
requests==2.32.3
python-docx==1.1.0
PyPDF2==3.0.1
beautifulsoup4==4.12.3
aiohttp==3.9.3
aiosqlite==0.19.0
